+++ py/desispec/scripts/healpix_redshifts.py
@@ -139,7 +139,7 @@ def main(args):
             log.info(f"Dry run so not creating the batch script: {batchscript}"
                      + f"\tfor {healpixels=}, {args.survey=}, {args.program=}")
 
-        cmd = ['sbatch' ,]
+        cmd = ['sbatch' , '--kill-on-invalid-dep=yes']
         if args.batch_reservation:
             cmd.extend(['--reservation', args.batch_reservation])
         if args.batch_dependency: