+++ py/desispec/test/test_extract.py
@@ -1,14 +1,7 @@
 from __future__ import absolute_import, division, print_function
 
-try:
-    from specter.psf import load_psf
-    import gpu_specter
-    nospecter = False
-except ImportError:
-    from desiutil.log import get_logger
-    log = get_logger()
-    log.error('specter and/or gpu_specter not installed; skipping extraction tests')
-    nospecter = True
+from desispec.test.util import installed
+nospecter = not installed('specter', 'gpu_specter')
 
 import unittest
 import uuid
@@ -20,7 +13,7 @@ from importlib import resources
 
 import desispec.image
 import desispec.io
-import desispec.scripts.extract
+from desispec.test.util import installed
 
 from astropy.io import fits
 import numpy as np
@@ -29,6 +22,9 @@ class TestExtract(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
+        if nospecter:
+            return  # don't bother; can't extraction tests without specter
+
         cls.origdir = os.getcwd()
         cls.testdir = tempfile.mkdtemp()
         os.chdir(cls.testdir)
@@ -54,6 +50,9 @@ class TestExtract(unittest.TestCase):
         cls.img = img
 
     def setUp(self):
+        if nospecter:
+            self.skipTest('Extraction tests require specter and/or gpu_specter')
+
         os.chdir(self.testdir)
         for filename in (self.outfile, self.outmodel):
             if os.path.exists(filename):
@@ -62,13 +61,14 @@ class TestExtract(unittest.TestCase):
     @classmethod
     def tearDownClass(cls):
         #- Remove testdir only if it was created by tempfile.mkdtemp
-        if cls.testdir.startswith(tempfile.gettempdir()) and os.path.exists(cls.testdir):
+        if hasattr(cls, 'testdir') and cls.testdir.startswith(tempfile.gettempdir()) and os.path.exists(cls.testdir):
             shutil.rmtree(cls.testdir)
 
-        os.chdir(cls.origdir)
+        if hasattr(cls, 'origdir'):
+            os.chdir(cls.origdir)
 
-    @unittest.skipIf(nospecter, 'specter not installed; skipping extraction test')
     def test_extract(self):
+        import desispec.scripts.extract
         template = "desi_extract_spectra -i {} -p {} -w 7500,7600,0.75 -f {} -s 0 -n 5 --bundlesize 5 -o {} -m {}"
 
         cmd = template.format(self.imgfile, self.psffile, self.fibermapfile, self.outfile, self.outmodel)
@@ -125,6 +125,7 @@ class TestExtract(unittest.TestCase):
         """
         Compare specter and gpu_specter extractions
         """
+        import desispec.scripts.extract
         #- should also work with bundles and not starting at spectrum 0
         cmd = template.format(self.imgfile, self.psffile, self.fibermapfile,
                 self.outfile, self.outmodel, specmin, nspec)