+++ doc/changes.rst
@@ -2,7 +2,45 @@
 desispec Change Log
 ===================
 
-0.70.0 (unreleased)
+0.70.1 (unreleased)
+-------------------
+
+* Set ``$DESI_SPECTRO_ROBUST=True`` for preproc to proceed anyway
+  even without dark or bias models (PR `#2577`_).
+* Update _find_zeros to also use exptable PROGRAM (PR `#2578`_).
+* Set FIBERSTATUS=UNASSIGNED for targets with OBJTYPE != TGT or SKY
+  (PR `#2579`_).
+* ``assemble_fibermap`` set OBJTYPE=BAD for non-SKY non-TGT targets
+  (PR `#2580`_).
+* Improve robustness of ``desi_compute_psf`` and ``desi_average_psf`` when
+  individual bundles fail (PR `#2584`_).
+* For zcatalogs, make ``EFFTIME_SPEC`` program-dependent to match tile usage
+    (PR `#2583`_).
+* Add ``flagged_fibers.ecsv`` to flag fibers for specific exposures
+  (PR `#2588`_).
+* Add ``copy`` argument to Spectra object (PR `#2590`_).
+* Apply gains to preprocessed darks (PR `#2591`_).
+* Add ``FLAT_TO_PSF_FLUX`` to fibermap (PR `#2597`_).
+* Make ``FLAT_TO_PSF_FLUX`` and ``PSF_TO_FIBER_SPECFLUX`` float32 (PR `#2598`_).
+* Add detailed control over ordering in zall files (PR `#2568`_).
+* Add tile info in tile QA (PR `#2600`_).
+
+.. _`#2568`: https://github.com/desihub/desispec/pull/2568
+.. _`#2577`: https://github.com/desihub/desispec/pull/2577
+.. _`#2578`: https://github.com/desihub/desispec/pull/2578
+.. _`#2579`: https://github.com/desihub/desispec/pull/2579
+.. _`#2580`: https://github.com/desihub/desispec/pull/2580
+.. _`#2583`: https://github.com/desihub/desispec/pull/2583
+.. _`#2584`: https://github.com/desihub/desispec/pull/2584
+.. _`#2588`: https://github.com/desihub/desispec/pull/2588
+.. _`#2590`: https://github.com/desihub/desispec/pull/2590
+.. _`#2591`: https://github.com/desihub/desispec/pull/2591
+.. _`#2597`: https://github.com/desihub/desispec/pull/2597
+.. _`#2598`: https://github.com/desihub/desispec/pull/2598
+.. _`#2600`: https://github.com/desihub/desispec/pull/2600
+
+
+0.70.0 (2025-12-04)
 -------------------
 
 Algorithms / core features:
@@ -17,6 +55,10 @@ Algorithms / core features:
 * Fix ``read_spectra`` indexing error (PR `#2508`_).
 * Magnitude calculation by photons not flux (PR `#2515`_).
 * Adjust cosmics parameters for red ITL CCDs (PR `#2557`_).
+* New ``plot_preproc`` script to plot specific tile/fiber/wavelength
+  thumbnails (PR `#2570`_).
+* Add ``desi_compute_psf --mpi`` option for MPI parallelism from the
+  command line (PR `#2576`_).
 
 Pipeline:
 
@@ -40,6 +82,7 @@ Pipeline:
 * Fix readonly write when no sky fibers (PR `#2523`_).
 * Purge dark_preproc in desi_purge_night (PR `#2553`_).
 * Make compute_dark crashes more explicit (PR `#2554`_).
+* Require minimum 6 darks instead of 4 for nightly dark (PR `#2575`_).
 
 Cleanup:
 
@@ -69,6 +112,7 @@ Cleanup:
 * Fix zall blank string bug (PR `#2566`_).
 * Add pyproject.toml and update top-level files (PR `#2569`_).
 * Fix bug in zall zcatlog PROGRAM and SURVEY header keywords (PR `#2572`_).
+* Installation and testing updates in prep for PyPI upload (PR `#2573`_).
 
 .. _`#2531`: https://github.com/desihub/desispec/pull/2531
 .. _`#2502`: https://github.com/desihub/desispec/pull/2502
@@ -127,7 +171,11 @@ Cleanup:
 .. _`#2562`: https://github.com/desihub/desispec/pull/2562
 .. _`#2566`: https://github.com/desihub/desispec/pull/2566
 .. _`#2569`: https://github.com/desihub/desispec/pull/2569
+.. _`#2570`: https://github.com/desihub/desispec/pull/2570
 .. _`#2572`: https://github.com/desihub/desispec/pull/2572
+.. _`#2573`: https://github.com/desihub/desispec/pull/2573
+.. _`#2575`: https://github.com/desihub/desispec/pull/2575
+.. _`#2576`: https://github.com/desihub/desispec/pull/2576
 
 
 0.69.0 (2025-03-14)